Ignore this script, sudo watchdog concept
This script just contains enough of a proof of concept for a watchdog script that keeps the sudo service alive, so it seems thus far.
Needs to be adjusted to run for a couple hours to see if the two threads are able to see the sudo session, or if like the variables
attempt I tried, it just fails due to context or something. but it keeps alive for 16 minutes.
So, hopefully.